koRpus.lang.en-package {koRpus.lang.en}	R Documentation

Language Support for 'koRpus' Package: English

Description

Adds support for the English language to the 'koRpus' package. To ask for help, report bugs, suggest feature improvements, or discuss the global development of the package, please consider subscribing to the koRpus-dev mailing list (<https://korpusml.reaktanz.de>).

Details

The DESCRIPTION file:

Package: 	 koRpus.lang.en

Type: 	 Package

Version: 	 0.1-4

Date: 	 2020-10-24

Depends: 	 R (>= 3.1),koRpus (>= 0.11-2)

Encoding: 	 UTF-8

License: 	 GPL (>= 3)

LazyLoad: 	 yes

URL: 	 https://reaktanz.de/?c=hacking&s=koRpus

Author(s)

NA

Maintainer: NA

See Also

Useful links:

https://reaktanz.de/?c=hacking&s=koRpus

Report bugs at https://github.com/unDocUMeantIt/koRpus.lang.en/issues

[Package koRpus.lang.en version 0.1-4 Index]